renaming files to resolve conflict in staging
