+++ src/App.js
@@ -5,7 +5,7 @@ import useLocalStorageState from 'use-local-storage-state'
 import { Header } from './components/Header'
 import { NavBar } from './components/NavBar'
 import { Login } from './components/Login'
-import { CardSnapshot } from './components/CardSnapshot'
+import { CardSnapshot } from './components/CardSnapshot-simple'
 
 function App() {
   const [token, setToken] = useLocalStorageState('token', null)
